# specs/001-erp-core/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: WellKorea ERP API
  description: |
    REST API for WellKorea Integrated Work System.
    Manages Projects (identified by JobCode), quotations, approvals, production tracking, deliveries, invoicing, and payments.

    **Key Concepts**:
    - **Project**: Core domain entity representing a customer work request
    - **JobCode**: Unique business identifier for each Project (format: WK2{year}-{sequence}-{date})
    - **Approval**: Separate domain handling approval workflows for quotations and other entities

    ## CQRS Pattern

    This API follows the **Command Query Responsibility Segregation (CQRS)** pattern:

    ### Command Endpoints (Write Operations)
    Command endpoints (POST for create/submit, PUT for update) return a minimal `CommandResult`:
    ```json
    {
      "id": 123,
      "message": "Quotation created successfully"
    }
    ```
    After a command succeeds, clients should fetch the updated entity via the corresponding query endpoint if they need fresh data.

    ### Query Endpoints (Read Operations)
    Query endpoints (GET) return full entity representations with all fields.

    ### Benefits
    - **Simplicity**: Commands don't need to assemble complex response objects
    - **Consistency**: Clients always get the latest state from query endpoints
    - **Scalability**: Commands and queries can be scaled independently
    - **Event-driven**: Commands can publish domain events for side effects (e.g., email notifications)
...
servers:
...
paths:
...
  /admin/products:
    post:
      tags:
        - Admin
      summary: Create product in catalog
      operationId: createProduct
      security:
        - BearerAuth: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProductRequest'
      responses:
        '201':
          description: Product created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Only Admin can manage product catalog

    get:
      tags:
        - Admin
      summary: List products in catalog
      operationId: listProducts
      security:
        - BearerAuth: [ ]
      parameters:
        - name: product_type_id
          in: query
          schema:
            type: string
            format: uuid
        - name: search
          in: query
          description: Search by product name or SKU
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: List of products
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: array
                    items:
                      $ref: '#/components/schemas/ProductResponse'
                  totalElements:
                    type: integer
                  totalPages:
                    type: integer
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
...
    ProductResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        sku:
          type: string
        name:
          type: string
        description:
          type: string
        product_type_id:
          type: string
          format: uuid
        base_unit_price:
          type: number
          format: double
        unit:
          type: string
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
    CreateProductRequest:
      type: object
      required:
        - sku
        - name
        - product_type_id
      properties:
        sku:
          type: string
          maxLength: 50
        name:
          type: string
          maxLength: 255
        description:
          type: string
        product_type_id:
          type: string
          format: uuid
        base_unit_price:
          type: number
          format: double
        unit:
          type: string

    # ========================================================================
    # User Schemas
    # ========================================================================
...
  responses:
    BadRequest:
      description: Invalid request payload
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
              details:
                type: array
                items:
                  type: object
                  properties:
                    field:
                      type: string
                    message:
                      type: string
    Unauthorized:
      description: Missing or invalid JWT token
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: Unauthorized
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - BearerAuth: [ ]